Revert "Reverse thumbs like endgame"
I tried it for about an hour. I can't say with certainty that the "this
feels stretched out to me" is nature or nuture. I have a possibly fake
memory that the corne felt weird when I first started using it, but
anyway.

I think in general that I prefer my thumb curled and roughly aligned
with the outside of my pointer finger.

This reverts commit 67dcd8ec730f65214151b6242cf7404dace9ddb9.